obj: move palloc from pmalloc.c to palloc.c
