+++ src/test/obj_constructor/memcheck1.log.match
@@ -7,7 +7,7 @@
 ==$(N)== Invalid write of size 4
 ==$(N)==    at 0x$(X): main (obj_constructor.c:$(N))
 ==$(N)==  Address 0x$(X) is $(N) bytes inside a block of size $(N) free'd
-==$(N)==    at 0x$(X): palloc_operation (pmalloc.c:$(N))
+==$(N)==    at 0x$(X): palloc_operation (palloc.c:$(N))
 ==$(N)==    by 0x$(X): obj_free (obj.c:$(N))
 ==$(N)==    by 0x$(X): pmemobj_free (obj.c:$(N))
 ==$(N)==    by 0x$(X): main (obj_constructor.c:$(N))
